Добавление Environment
Добавлен файл environment, все ссылки на сервер заменены глобальными переменными.

+++ Remonty/wells-front/src/app/components/vidRemonta/vid/vid.component.ts
@@ -3,6 +3,9 @@ import { HttpClient } from '@angular/common/http';
 import { ServiceService } from 'src/app/Service/service.service';
 import { Router } from '@angular/router';
 import { Vid } from './Vid';
+import { environment } from 'src/environments/environment';
+import { Detail } from 'src/app/components/details/detail/Detail';
+
 
 @Component({
   selector: 'app-vid',
@@ -12,11 +15,11 @@ import { Vid } from './Vid';
 export class VidComponent {
     title = 'vid-angular-project';
     vidy!: Vid[];
-    i = 0;
+    details!:Detail[];
+    i = 0; k = 0;
 
     constructor(private httpClient: HttpClient, private router:Router, private service:ServiceService) {
-        httpClient.get<Vid[]>("http://localhost:8081/vidy")
-        //httpClient.get<Vid[]>("http://192.168.7.111:8081/vidy")
+        httpClient.get<Vid[]>(environment.UrlVid)
         .subscribe(result=>{
             this.vidy = result;
         })
@@ -40,11 +43,21 @@ export class VidComponent {
     }
 
     Delete(vid:Vid){
-        this.service.deleteVid(vid)
-        .subscribe(data =>{
-            this.vidy=this.vidy.filter(r => r !== vid );
-            this.router.navigate(["vidyList"]);
-            alert("Запись удалена!");
+        this.httpClient.get<Detail[]>(environment.UrlDetail)
+            .subscribe(result=>{
+            this.details = result;
+            this.k = 0;
+            for (let detail of this.details) {
+                if (vid.name == Object.values(detail)[0] && 0 == Object.values(detail)[1]) { this.k = 1; }
+            }
+            if (this.k == 1)
+            this.service.deleteVid(vid)
+            .subscribe(data =>{
+                this.vidy=this.vidy.filter(r => r !== vid );
+                this.router.navigate(["vidyList"]);
+                alert("Запись удалена!");
+            })
+            else alert("Невозможно удалить!\nРешение: удалите или измените информацию в списке ремонтов.");
         })
     }
 }
